Support for synchronizing assets onto a DevFS
