+++ packages/flutter_tools/test/devfs_test.dart
@@ -4,6 +4,7 @@
 
 import 'dart:io';
 
+import 'package:flutter_tools/src/asset.dart';
 import 'package:flutter_tools/src/devfs.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
@@ -18,6 +19,8 @@ void main() {
   String basePath;
   MockDevFSOperations devFSOperations = new MockDevFSOperations();
   DevFS devFS;
+  AssetBundle assetBundle = new AssetBundle();
+  assetBundle.entries.add(new AssetBundleEntry.fromString('a.txt', ''));
   group('devfs', () {
     testUsingContext('create local file system', () async {
       tempDir = Directory.systemTemp.createTempSync();
@@ -38,8 +41,6 @@ void main() {
     testUsingContext('modify existing file on local file system', () async {
       File file = new File(path.join(basePath, filePath));
       file.writeAsBytesSync(<int>[1, 2, 3, 4, 5, 6]);
-    });
-    testUsingContext('update dev file system', () async {
       await devFS.update();
       expect(devFSOperations.contains('writeFile test bar/foo.txt'), isTrue);
     });
@@ -47,10 +48,28 @@ void main() {
       File file = new File(path.join(basePath, filePath2));
       await file.parent.create(recursive: true);
       file.writeAsBytesSync(<int>[1, 2, 3, 4, 5, 6, 7]);
-    });
-    testUsingContext('update dev file system', () async {
       await devFS.update();
       expect(devFSOperations.contains('writeFile test foo/bar.txt'), isTrue);
+    });
+    testUsingContext('delete a file from the local file system', () async {
+      File file = new File(path.join(basePath, filePath));
+      await file.delete();
+      await devFS.update();
+      expect(devFSOperations.contains('deleteFile test bar/foo.txt'), isTrue);
+    });
+    testUsingContext('add file in an asset bundle', () async {
+      await devFS.update(assetBundle);
+      expect(devFSOperations.contains('writeFile test build/flx/a.txt'), isTrue);
+    });
+    testUsingContext('add a file to the asset bundle', () async {
+      assetBundle.entries.add(new AssetBundleEntry.fromString('b.txt', ''));
+      await devFS.update(assetBundle);
+      expect(devFSOperations.contains('writeFile test build/flx/b.txt'), isTrue);
+    });
+    testUsingContext('delete a file from the asset bundle', () async {
+      assetBundle.entries.clear();
+      await devFS.update(assetBundle);
+      expect(devFSOperations.contains('deleteFile test build/flx/b.txt'), isTrue);
     });
     testUsingContext('delete dev file system', () async {
       await devFS.destroy();
